Generate keyword documentation without a kubernetes cluster
For a Docker image with Robotframework and additional
libraries we generate the keyword documentation for all
included Robot libraries with robotframework-hub-static
and robotframework-hub-bli.

For this a generic way to run libdoc is necessary and
library specific initialization would break the generic
approach.

So far robotframework-kubelibrary was the only library
where the generation of keyword documentation failed
and required access to a Kubernetes cluster.

This patch evaluates environment variable INIT_FOR_LIBDOC_ONLY.
When this variable is set to 1, the keyword documentation can
be created without the need to access a cluster.